refactor: convert to JS
