+++ src/lib/get-slides-from-content.js
@@ -1,0 +1,7 @@
+function getSlidesFromContent(content) {
+	return content.split("---");
+}
+
+module.exports = {
+	getSlidesFromContent,
+};
