--- src/lib/get-slides-from-content.ts
@@ -1,3 +1,0 @@
-export function getSlidesFromContent(content: string): string[] {
-	return content.split("---");
-}
